+++ src/ooo_cpu.cc
@@ -93,7 +93,7 @@ void O3_CPU::init_instruction(ooo_model_instr arch_instr)
     }
   }
 
-  for (int i = 0; i < NUM_INSTR_SOURCES; i++) {
+  for (size_t i = 0; i < NUM_INSTR_SOURCES; i++) {
     switch (arch_instr.source_registers[i]) {
     case 0:
       break;
@@ -243,7 +243,7 @@ void O3_CPU::init_instruction(ooo_model_instr arch_instr)
   // fast warmup eliminates register dependencies between instructions
   // branch predictor, cache contents, and prefetchers are still warmed up
   if (!warmup_complete[cpu]) {
-    for (int i = 0; i < NUM_INSTR_SOURCES; i++) {
+    for (size_t i = 0; i < NUM_INSTR_SOURCES; i++) {
       arch_instr.source_registers[i] = 0;
     }
     for (uint32_t i = 0; i < MAX_INSTR_DESTINATIONS; i++) {
